+++ src/pricingengines/vanilla/juquadraticengine.cpp
@@ -1,0 +1,31 @@
+/*
+ * PyQuantLib: Python bindings for QuantLib
+ * https://github.com/quantales/pyquantlib
+ *
+ * Copyright (c) 2025 Yassine Idyiahia
+ * SPDX-License-Identifier: BSD-3-Clause
+ * See LICENSE for details.
+ *
+ * ---
+ * QuantLib is Copyright (c) 2000-2025 The QuantLib Authors
+ * https://www.quantlib.org/
+ */
+
+#include "pyquantlib/pyquantlib.h"
+#include <ql/pricingengines/vanilla/juquadraticengine.hpp>
+#include <ql/processes/blackscholesprocess.hpp>
+#include <pybind11/pybind11.h>
+
+namespace py = pybind11;
+using namespace QuantLib;
+
+void ql_pricingengines::juquadraticengine(py::module_& m) {
+    py::class_<JuQuadraticApproximationEngine,
+               ext::shared_ptr<JuQuadraticApproximationEngine>,
+               PricingEngine>(
+        m, "JuQuadraticApproximationEngine",
+        "Ju quadratic approximation for American options.")
+        .def(py::init<ext::shared_ptr<GeneralizedBlackScholesProcess>>(),
+            py::arg("process"),
+            "Constructs Ju quadratic approximation engine.");
+}